Merge and Append your Postural data

This function merges all .txt files directly exported from the AMTI NetForce software. It re-names columns (un-named by the AMTI software), compute automatically CoP on both the antero-posterior (CoP-Y) and medio-lateral (CoP-X) axes, and add the time course during the session to facilitate further data management. Finally, it optionnally saves the resulting output in your file directory. Note that the SessionDuaration argument is optionnal and only serves the purpose of verifying data integrity: if you have sessions of different length, you should not use it. However if your sessions are supposed to be of equal length, then it might make sense to use the argument in order to assess whether the number of rows expected from Sample rate and Session duration match the Actual duration of the sessions.

Usage

Merge_PosData(
  directory_path,
  SampleRate,
  SessionDuration = NULL,
  write_csv = FALSE
)

Arguments

directory_path

directory to a folder containing all the .txt outputs from AMTI Netforce that you wish to merge in a single data set.

SampleRate

Sample rate used in your protocol

SessionDuration

Duration of your protocol (optionnal)

write_csv

Logical. If TRUE, the merged data is saved as a CSV file in the same directory. Default is FALSE.

Value

a dataframe containing all merged postural files, with additional columns: CoP-Y, CoP-X, time course, and file name.

Examples

#Find subdirectory of Example data in the original .txt format exported from AMTI Netforce software
# Note: we need to convert compressed rdata to original txt file
files <- paste0("Postural_Data", LETTERS[1:6])

# Locate the directory containing the .Rdata files within the package
data_dir <- system.file("data", package = "BalanceMate")

# Create a temporary directory to store the .txt files
temp_data_dir <- file.path(tempdir(), "data")
dir.create(temp_data_dir, showWarnings = FALSE)

# Process each file: load, optionally add a blank row, and write to .txt
invisible(lapply(files, function(f) {
  # Load the .Rdata file from the package's extdata directory
  load(file.path(data_dir, paste0(f, ".Rdata")))

  data <- get(f)
  # Write the data to a .txt file in the temporary directory
  write.table(data, file = file.path(temp_data_dir, paste0(f, ".txt")), sep = ",",
  row.names = FALSE, col.names = FALSE, quote = FALSE)
}))

Data <- Merge_PosData(temp_data_dir, SampleRate = 100, SessionDuration = 100)