Postural data - Block B - txt file

A .txt file version of a full block of postural measurement during a session in the Monéger et al. (submitted) experiment. Block duration is 331s, and Sample rate was 100Hz. As such, each block corresponds to a data frame of 33,100 rows with the 6 exported columns from AMTI NetForce, unconveniently un-named. They correspond to Mx, My, Mz, Fx, Fy, Fz. This is a plain text version of the postural data for Block B. It is stored as a .txt file in the inst/extdata directory of the BalanceMate package.

Usage

```r
# To load the data from the text file:
path_to_data <- system.file("extdata", "Postural_DataB.txt", package = "BalanceMate")
Postural_DataB_txt <- read.table(path_to_data, header = TRUE, sep = "\t")
```

Format

Postural_DataB

A data frame with 33,100 rows and 6 columns:

Mx

Un-named column corresponding to Mx

My

Un-named column corresponding to My

Mz

Un-named column corresponding to Mz

Fx

Un-named column corresponding to Fx

Fy

Un-named column corresponding to Fy

Fz

Un-named column corresponding to Fz

Source

<Monéger, J., Besche-Richard, C., Chaby, L., & Vergilino-Perez D. (Submitted in Emotion). Should I Stay or Should I go: An Exploration of Spontaneous Postural Behaviours Following Threatening Emotion Perception.>

Examples

path_to_data <- system.file("extdata", "Postural_DataB.txt", package = "BalanceMate")
Postural_DataB <- read.table(path_to_data, header = TRUE, sep = "\t")