Postural data - Block E (R data)

A shortened block of postural measurement during a session in the Monéger et al. (submitted) experiment. Block duration is 100s (originally 331s), and Sample rate was 100Hz. As such, each block corresponds to a data frame of 10,000 rows with the 6 exported columns from AMTI NetForce, unconveniently un-named. They correspond to Mx, My, Mz, Fx, Fy, Fz.

Usage

data(Postural_DataE)

Format

Postural_DataE

A data frame with 10,000 rows and 6 columns:

Mx

Un-named column corresponding to Mx- moment around the X axis

My

Un-named column corresponding to My- moment around the Y axis

Mz

Un-named column corresponding to Mz- moment around the Z axis

Fx

Un-named column corresponding to Fx- force in the X axis

Fy

Un-named column corresponding to Fy- force in the Y axis

Fz

Un-named column corresponding to Fz- force in the Z axis

Source

<Monéger, J., Besche-Richard, C., Chaby, L., & Vergilino-Perez D. (Submitted in Emotion). Should I Stay or Should I go: An Exploration of Spontaneous Postural Behaviours Following Threatening Emotion Perception.>