Compute mean Sway Path Length

Compute mean Sway Path Length

Usage

SPL_ComputeR(
  data,
  CoPX_col,
  CoPY_col,
  ID,
  time_col = NULL,
  epoch_length = NULL
)

Arguments

data

a data frame

CoPX_col

a column in the data frame storing the CoP-X data

CoPY_col

a column in the data frame storing the CoP-Y data

ID

a column in the data frame storing Participants' unique identifiers

time_col

a column in the data frame storing Time course

epoch_length

numeric value: epoch duration over which you want to compute the sway path length (in seconds)

Value

a synthetised data frame (i.e., shorter) with an additional column (Sway path lenght)

Details

If the specified epoch is not a multiple of the protocol duration (e.g., user inputs an epoch of 3s for a protocol of 10s), then the function will compute the SPL over the unbalanced time periods (in the latter example, 4 mean SPL would be computed with the three first periods corresponding to 3s intervals, but the last one corresponding to the remaining time – that is 1s interval). A warning will be issued to inform user about unbalanced epoch computations.

Examples

# Find subdirectory of Example data in the original .txt format exported from AMTI Netforce software
# Note: we need to convert compressed rdata to original txt file
files <- paste0("Postural_Data", LETTERS[1:6])

# Locate the directory containing the .Rdata files within the package
data_dir <- system.file("data", package = "BalanceMate")

# Create a temporary directory to store the .txt files
temp_data_dir <- file.path(tempdir(), "data")
dir.create(temp_data_dir, showWarnings = FALSE)

# Process each file: load, optionally add a blank row, and write to .txt
invisible(lapply(files, function(f) {
  # Load the .Rdata file from the package's extdata directory
  load(file.path(data_dir, paste0(f, ".Rdata")))

  data <- get(f)
  # Write the data to a .txt file in the temporary directory
  write.table(data, file = file.path(temp_data_dir, paste0(f, ".txt")), sep = ",",
  row.names = FALSE, col.names = FALSE, quote = FALSE)
}))

Data <- Merge_PosData(temp_data_dir, SampleRate = 100, SessionDuration = 100)

SPL_ComputeR(data = Data,
    CoPX_col = "CoP_X",
    CoPY_col = "CoP_Y",
    ID = "file_name",
    time_col = "Time",
    epoch_length = 1)
#>                        participant_id epoch sway_path_length
#> Postural_DataA.txt.1   Postural_DataA     1        1.5983162
#> Postural_DataA.txt.2   Postural_DataA     2        1.5800963
#> Postural_DataA.txt.3   Postural_DataA     3        2.5338072
#> Postural_DataA.txt.4   Postural_DataA     4        1.5498590
#> Postural_DataA.txt.5   Postural_DataA     5        2.0116923
#> Postural_DataA.txt.6   Postural_DataA     6        1.3294412
#> Postural_DataA.txt.7   Postural_DataA     7        1.8442789
#> Postural_DataA.txt.8   Postural_DataA     8        1.8172048
#> Postural_DataA.txt.9   Postural_DataA     9        2.1942489
#> Postural_DataA.txt.10  Postural_DataA    10        4.5943216
#> Postural_DataA.txt.11  Postural_DataA    11        2.7838749
#> Postural_DataA.txt.12  Postural_DataA    12        1.9878485
#> Postural_DataA.txt.13  Postural_DataA    13        2.2683356
#> Postural_DataA.txt.14  Postural_DataA    14        1.4977833
#> Postural_DataA.txt.15  Postural_DataA    15        1.3812698
#> Postural_DataA.txt.16  Postural_DataA    16        1.4336251
#> Postural_DataA.txt.17  Postural_DataA    17        1.4987895
#> Postural_DataA.txt.18  Postural_DataA    18        1.5486270
#> Postural_DataA.txt.19  Postural_DataA    19        1.2878721
#> Postural_DataA.txt.20  Postural_DataA    20        1.1743384
#> Postural_DataA.txt.21  Postural_DataA    21        1.5910956
#> Postural_DataA.txt.22  Postural_DataA    22        5.7486429
#> Postural_DataA.txt.23  Postural_DataA    23        2.9268332
#> Postural_DataA.txt.24  Postural_DataA    24        1.8025851
#> Postural_DataA.txt.25  Postural_DataA    25        1.3128612
#> Postural_DataA.txt.26  Postural_DataA    26        2.7313395
#> Postural_DataA.txt.27  Postural_DataA    27        1.8423063
#> Postural_DataA.txt.28  Postural_DataA    28        1.7787609
#> Postural_DataA.txt.29  Postural_DataA    29        1.3306148
#> Postural_DataA.txt.30  Postural_DataA    30        1.4743078
#> Postural_DataA.txt.31  Postural_DataA    31        1.2186638
#> Postural_DataA.txt.32  Postural_DataA    32        2.0590614
#> Postural_DataA.txt.33  Postural_DataA    33        1.6304480
#> Postural_DataA.txt.34  Postural_DataA    34        1.2751088
#> Postural_DataA.txt.35  Postural_DataA    35        1.3963437
#> Postural_DataA.txt.36  Postural_DataA    36        1.2061783
#> Postural_DataA.txt.37  Postural_DataA    37        1.6585383
#> Postural_DataA.txt.38  Postural_DataA    38        1.3075318
#> Postural_DataA.txt.39  Postural_DataA    39        1.3223952
#> Postural_DataA.txt.40  Postural_DataA    40        1.8913456
#> Postural_DataA.txt.41  Postural_DataA    41        2.3011455
#> Postural_DataA.txt.42  Postural_DataA    42        1.9323457
#> Postural_DataA.txt.43  Postural_DataA    43        1.6608511
#> Postural_DataA.txt.44  Postural_DataA    44        1.4891183
#> Postural_DataA.txt.45  Postural_DataA    45        1.7434736
#> Postural_DataA.txt.46  Postural_DataA    46        1.3509021
#> Postural_DataA.txt.47  Postural_DataA    47        1.6274212
#> Postural_DataA.txt.48  Postural_DataA    48        1.3194831
#> Postural_DataA.txt.49  Postural_DataA    49        1.0409888
#> Postural_DataA.txt.50  Postural_DataA    50        1.3361560
#> Postural_DataA.txt.51  Postural_DataA    51        1.2876868
#> Postural_DataA.txt.52  Postural_DataA    52        1.5713767
#> Postural_DataA.txt.53  Postural_DataA    53        2.9750884
#> Postural_DataA.txt.54  Postural_DataA    54        2.5554548
#> Postural_DataA.txt.55  Postural_DataA    55        2.5870491
#> Postural_DataA.txt.56  Postural_DataA    56        2.3674357
#> Postural_DataA.txt.57  Postural_DataA    57        2.0149878
#> Postural_DataA.txt.58  Postural_DataA    58        1.4395540
#> Postural_DataA.txt.59  Postural_DataA    59        2.5909728
#> Postural_DataA.txt.60  Postural_DataA    60        1.4450378
#> Postural_DataA.txt.61  Postural_DataA    61        1.3481751
#> Postural_DataA.txt.62  Postural_DataA    62        1.4464002
#> Postural_DataA.txt.63  Postural_DataA    63        2.5995735
#> Postural_DataA.txt.64  Postural_DataA    64        1.8513490
#> Postural_DataA.txt.65  Postural_DataA    65        1.5170661
#> Postural_DataA.txt.66  Postural_DataA    66        1.2699765
#> Postural_DataA.txt.67  Postural_DataA    67        1.0810265
#> Postural_DataA.txt.68  Postural_DataA    68        1.0436240
#> Postural_DataA.txt.69  Postural_DataA    69        1.7440447
#> Postural_DataA.txt.70  Postural_DataA    70        1.6605288
#> Postural_DataA.txt.71  Postural_DataA    71        1.5540605
#> Postural_DataA.txt.72  Postural_DataA    72        1.5481394
#> Postural_DataA.txt.73  Postural_DataA    73        1.5890152
#> Postural_DataA.txt.74  Postural_DataA    74        1.1256542
#> Postural_DataA.txt.75  Postural_DataA    75        1.4931746
#> Postural_DataA.txt.76  Postural_DataA    76        1.2860192
#> Postural_DataA.txt.77  Postural_DataA    77        1.3725983
#> Postural_DataA.txt.78  Postural_DataA    78        1.1973834
#> Postural_DataA.txt.79  Postural_DataA    79        1.2398819
#> Postural_DataA.txt.80  Postural_DataA    80        1.7671620
#> Postural_DataA.txt.81  Postural_DataA    81        1.5839288
#> Postural_DataA.txt.82  Postural_DataA    82        1.5663982
#> Postural_DataA.txt.83  Postural_DataA    83        2.5205773
#> Postural_DataA.txt.84  Postural_DataA    84        1.8863628
#> Postural_DataA.txt.85  Postural_DataA    85        1.5616868
#> Postural_DataA.txt.86  Postural_DataA    86        2.1023681
#> Postural_DataA.txt.87  Postural_DataA    87        1.5205284
#> Postural_DataA.txt.88  Postural_DataA    88        1.4060439
#> Postural_DataA.txt.89  Postural_DataA    89        1.1769761
#> Postural_DataA.txt.90  Postural_DataA    90        1.3322437
#> Postural_DataA.txt.91  Postural_DataA    91        1.1166559
#> Postural_DataA.txt.92  Postural_DataA    92        2.4652289
#> Postural_DataA.txt.93  Postural_DataA    93        1.1869255
#> Postural_DataA.txt.94  Postural_DataA    94        1.7026789
#> Postural_DataA.txt.95  Postural_DataA    95        1.3068029
#> Postural_DataA.txt.96  Postural_DataA    96        1.0486204
#> Postural_DataA.txt.97  Postural_DataA    97        1.3922058
#> Postural_DataA.txt.98  Postural_DataA    98        1.5641587
#> Postural_DataA.txt.99  Postural_DataA    99        1.4050869
#> Postural_DataA.txt.100 Postural_DataA   100        2.3916049
#> Postural_DataB.txt.1   Postural_DataB     1        1.9993064
#> Postural_DataB.txt.2   Postural_DataB     2        3.6149971
#> Postural_DataB.txt.3   Postural_DataB     3        1.8997802
#> Postural_DataB.txt.4   Postural_DataB     4        3.7049778
#> Postural_DataB.txt.5   Postural_DataB     5        2.7692700
#> Postural_DataB.txt.6   Postural_DataB     6        2.7828477
#> Postural_DataB.txt.7   Postural_DataB     7        2.1662586
#> Postural_DataB.txt.8   Postural_DataB     8        1.6725051
#> Postural_DataB.txt.9   Postural_DataB     9        4.4640760
#> Postural_DataB.txt.10  Postural_DataB    10        7.6218184
#> Postural_DataB.txt.11  Postural_DataB    11        5.1891701
#> Postural_DataB.txt.12  Postural_DataB    12        2.3913982
#> Postural_DataB.txt.13  Postural_DataB    13        1.2667835
#> Postural_DataB.txt.14  Postural_DataB    14        1.4880953
#> Postural_DataB.txt.15  Postural_DataB    15        1.3263749
#> Postural_DataB.txt.16  Postural_DataB    16        1.5252706
#> Postural_DataB.txt.17  Postural_DataB    17        1.6193432
#> Postural_DataB.txt.18  Postural_DataB    18        1.8249405
#> Postural_DataB.txt.19  Postural_DataB    19        1.5042515
#> Postural_DataB.txt.20  Postural_DataB    20        1.2194327
#> Postural_DataB.txt.21  Postural_DataB    21        1.3409677
#> Postural_DataB.txt.22  Postural_DataB    22        1.2182704
#> Postural_DataB.txt.23  Postural_DataB    23        1.2441254
#> Postural_DataB.txt.24  Postural_DataB    24        1.0805648
#> Postural_DataB.txt.25  Postural_DataB    25        1.2110604
#> Postural_DataB.txt.26  Postural_DataB    26        1.0304113
#> Postural_DataB.txt.27  Postural_DataB    27        1.1884130
#> Postural_DataB.txt.28  Postural_DataB    28        1.2800334
#> Postural_DataB.txt.29  Postural_DataB    29        2.0809713
#> Postural_DataB.txt.30  Postural_DataB    30        1.3510501
#> Postural_DataB.txt.31  Postural_DataB    31        1.4131942
#> Postural_DataB.txt.32  Postural_DataB    32        1.3172170
#> Postural_DataB.txt.33  Postural_DataB    33        2.2459072
#> Postural_DataB.txt.34  Postural_DataB    34        3.0237646
#> Postural_DataB.txt.35  Postural_DataB    35        2.8126739
#> Postural_DataB.txt.36  Postural_DataB    36        3.5869274
#> Postural_DataB.txt.37  Postural_DataB    37        1.7076560
#> Postural_DataB.txt.38  Postural_DataB    38        1.1974150
#> Postural_DataB.txt.39  Postural_DataB    39        1.1865362
#> Postural_DataB.txt.40  Postural_DataB    40        1.0716129
#> Postural_DataB.txt.41  Postural_DataB    41        1.1664721
#> Postural_DataB.txt.42  Postural_DataB    42        1.1701273
#> Postural_DataB.txt.43  Postural_DataB    43        1.1544756
#> Postural_DataB.txt.44  Postural_DataB    44        1.2666190
#> Postural_DataB.txt.45  Postural_DataB    45        1.1877100
#> Postural_DataB.txt.46  Postural_DataB    46        1.1950169
#> Postural_DataB.txt.47  Postural_DataB    47        1.0765019
#> Postural_DataB.txt.48  Postural_DataB    48        1.1807372
#> Postural_DataB.txt.49  Postural_DataB    49        1.2584924
#> Postural_DataB.txt.50  Postural_DataB    50        1.2621454
#> Postural_DataB.txt.51  Postural_DataB    51        1.4134676
#> Postural_DataB.txt.52  Postural_DataB    52        1.2776810
#> Postural_DataB.txt.53  Postural_DataB    53        1.1890717
#> Postural_DataB.txt.54  Postural_DataB    54        1.1623458
#> Postural_DataB.txt.55  Postural_DataB    55        1.2953379
#> Postural_DataB.txt.56  Postural_DataB    56        1.6175395
#> Postural_DataB.txt.57  Postural_DataB    57        1.1963502
#> Postural_DataB.txt.58  Postural_DataB    58        1.4880377
#> Postural_DataB.txt.59  Postural_DataB    59        1.4367873
#> Postural_DataB.txt.60  Postural_DataB    60        1.2845209
#> Postural_DataB.txt.61  Postural_DataB    61        1.3199869
#> Postural_DataB.txt.62  Postural_DataB    62        1.4453103
#> Postural_DataB.txt.63  Postural_DataB    63        2.6433306
#> Postural_DataB.txt.64  Postural_DataB    64        9.9300366
#> Postural_DataB.txt.65  Postural_DataB    65       11.9731710
#> Postural_DataB.txt.66  Postural_DataB    66        4.7574346
#> Postural_DataB.txt.67  Postural_DataB    67        3.3213717
#> Postural_DataB.txt.68  Postural_DataB    68        2.2582911
#> Postural_DataB.txt.69  Postural_DataB    69        2.3730972
#> Postural_DataB.txt.70  Postural_DataB    70        1.6937274
#> Postural_DataB.txt.71  Postural_DataB    71        2.2917937
#> Postural_DataB.txt.72  Postural_DataB    72        1.2253077
#> Postural_DataB.txt.73  Postural_DataB    73        1.1412332
#> Postural_DataB.txt.74  Postural_DataB    74        1.1311332
#> Postural_DataB.txt.75  Postural_DataB    75        1.1094702
#> Postural_DataB.txt.76  Postural_DataB    76        1.8353359
#> Postural_DataB.txt.77  Postural_DataB    77        2.4584632
#> Postural_DataB.txt.78  Postural_DataB    78        2.2893934
#> Postural_DataB.txt.79  Postural_DataB    79        1.9102396
#> Postural_DataB.txt.80  Postural_DataB    80        2.0565862
#> Postural_DataB.txt.81  Postural_DataB    81        1.8577065
#> Postural_DataB.txt.82  Postural_DataB    82        2.2652764
#> Postural_DataB.txt.83  Postural_DataB    83        3.3246068
#> Postural_DataB.txt.84  Postural_DataB    84        2.6623522
#> Postural_DataB.txt.85  Postural_DataB    85        2.5176935
#> Postural_DataB.txt.86  Postural_DataB    86        3.3843904
#> Postural_DataB.txt.87  Postural_DataB    87        1.6681593
#> Postural_DataB.txt.88  Postural_DataB    88        1.3698029
#> Postural_DataB.txt.89  Postural_DataB    89        3.9160274
#> Postural_DataB.txt.90  Postural_DataB    90        2.3103396
#> Postural_DataB.txt.91  Postural_DataB    91        1.4243455
#> Postural_DataB.txt.92  Postural_DataB    92        1.1395235
#> Postural_DataB.txt.93  Postural_DataB    93        1.2910717
#> Postural_DataB.txt.94  Postural_DataB    94        1.5361705
#> Postural_DataB.txt.95  Postural_DataB    95        1.4923331
#> Postural_DataB.txt.96  Postural_DataB    96        1.0911860
#> Postural_DataB.txt.97  Postural_DataB    97        1.9907082
#> Postural_DataB.txt.98  Postural_DataB    98        1.5503793
#> Postural_DataB.txt.99  Postural_DataB    99        1.1887538
#> Postural_DataB.txt.100 Postural_DataB   100        1.2729866
#> Postural_DataC.txt.1   Postural_DataC     1        2.1708935
#> Postural_DataC.txt.2   Postural_DataC     2        1.5178248
#> Postural_DataC.txt.3   Postural_DataC     3        1.8356737
#> Postural_DataC.txt.4   Postural_DataC     4        1.6416937
#> Postural_DataC.txt.5   Postural_DataC     5        1.4182885
#> Postural_DataC.txt.6   Postural_DataC     6        1.3499784
#> Postural_DataC.txt.7   Postural_DataC     7        1.2327053
#> Postural_DataC.txt.8   Postural_DataC     8        2.9223471
#> Postural_DataC.txt.9   Postural_DataC     9        2.1673513
#> Postural_DataC.txt.10  Postural_DataC    10        1.9244459
#> Postural_DataC.txt.11  Postural_DataC    11        1.9755487
#> Postural_DataC.txt.12  Postural_DataC    12        1.9125846
#> Postural_DataC.txt.13  Postural_DataC    13        1.3984779
#> Postural_DataC.txt.14  Postural_DataC    14        1.5726486
#> Postural_DataC.txt.15  Postural_DataC    15        1.9870694
#> Postural_DataC.txt.16  Postural_DataC    16        2.4752179
#> Postural_DataC.txt.17  Postural_DataC    17        3.0721594
#> Postural_DataC.txt.18  Postural_DataC    18        1.4073503
#> Postural_DataC.txt.19  Postural_DataC    19        1.1846433
#> Postural_DataC.txt.20  Postural_DataC    20        1.3571678
#> Postural_DataC.txt.21  Postural_DataC    21        1.5064339
#> Postural_DataC.txt.22  Postural_DataC    22        2.3345058
#> Postural_DataC.txt.23  Postural_DataC    23        1.9614723
#> Postural_DataC.txt.24  Postural_DataC    24        1.5877744
#> Postural_DataC.txt.25  Postural_DataC    25        1.7375864
#> Postural_DataC.txt.26  Postural_DataC    26        2.3008369
#> Postural_DataC.txt.27  Postural_DataC    27        2.7899198
#> Postural_DataC.txt.28  Postural_DataC    28        1.7258304
#> Postural_DataC.txt.29  Postural_DataC    29        2.1142499
#> Postural_DataC.txt.30  Postural_DataC    30        2.5922709
#> Postural_DataC.txt.31  Postural_DataC    31        1.4946086
#> Postural_DataC.txt.32  Postural_DataC    32        1.8592336
#> Postural_DataC.txt.33  Postural_DataC    33        0.9123813
#> Postural_DataC.txt.34  Postural_DataC    34        1.4499066
#> Postural_DataC.txt.35  Postural_DataC    35        2.0235025
#> Postural_DataC.txt.36  Postural_DataC    36        2.7988448
#> Postural_DataC.txt.37  Postural_DataC    37        9.0204642
#> Postural_DataC.txt.38  Postural_DataC    38       10.9159064
#> Postural_DataC.txt.39  Postural_DataC    39        6.3416559
#> Postural_DataC.txt.40  Postural_DataC    40        4.2698001
#> Postural_DataC.txt.41  Postural_DataC    41        3.3502309
#> Postural_DataC.txt.42  Postural_DataC    42        1.9513565
#> Postural_DataC.txt.43  Postural_DataC    43        2.0011734
#> Postural_DataC.txt.44  Postural_DataC    44        2.5938706
#> Postural_DataC.txt.45  Postural_DataC    45        1.6820098
#> Postural_DataC.txt.46  Postural_DataC    46        1.3650795
#> Postural_DataC.txt.47  Postural_DataC    47        1.3366357
#> Postural_DataC.txt.48  Postural_DataC    48        1.5864336
#> Postural_DataC.txt.49  Postural_DataC    49        1.6752675
#> Postural_DataC.txt.50  Postural_DataC    50        1.6003991
#> Postural_DataC.txt.51  Postural_DataC    51        2.1623369
#> Postural_DataC.txt.52  Postural_DataC    52        2.4851071
#> Postural_DataC.txt.53  Postural_DataC    53        1.5633574
#> Postural_DataC.txt.54  Postural_DataC    54        2.6062127
#> Postural_DataC.txt.55  Postural_DataC    55        2.2285088
#> Postural_DataC.txt.56  Postural_DataC    56        1.6531359
#> Postural_DataC.txt.57  Postural_DataC    57        1.4931481
#> Postural_DataC.txt.58  Postural_DataC    58        1.4525799
#> Postural_DataC.txt.59  Postural_DataC    59        2.8122426
#> Postural_DataC.txt.60  Postural_DataC    60        2.1490181
#> Postural_DataC.txt.61  Postural_DataC    61        1.8701064
#> Postural_DataC.txt.62  Postural_DataC    62        4.2065797
#> Postural_DataC.txt.63  Postural_DataC    63        2.0945988
#> Postural_DataC.txt.64  Postural_DataC    64        2.3066641
#> Postural_DataC.txt.65  Postural_DataC    65        1.9404705
#> Postural_DataC.txt.66  Postural_DataC    66        1.5529616
#> Postural_DataC.txt.67  Postural_DataC    67        1.2237019
#> Postural_DataC.txt.68  Postural_DataC    68        0.9652947
#> Postural_DataC.txt.69  Postural_DataC    69        1.5725934
#> Postural_DataC.txt.70  Postural_DataC    70        1.5916525
#> Postural_DataC.txt.71  Postural_DataC    71        1.6022134
#> Postural_DataC.txt.72  Postural_DataC    72        1.5433209
#> Postural_DataC.txt.73  Postural_DataC    73        1.2087788
#> Postural_DataC.txt.74  Postural_DataC    74        1.2982067
#> Postural_DataC.txt.75  Postural_DataC    75        1.2103939
#> Postural_DataC.txt.76  Postural_DataC    76        1.1390924
#> Postural_DataC.txt.77  Postural_DataC    77        1.5615221
#> Postural_DataC.txt.78  Postural_DataC    78        1.0120378
#> Postural_DataC.txt.79  Postural_DataC    79        1.2112522
#> Postural_DataC.txt.80  Postural_DataC    80        1.0922776
#> Postural_DataC.txt.81  Postural_DataC    81        1.4037880
#> Postural_DataC.txt.82  Postural_DataC    82        1.5122130
#> Postural_DataC.txt.83  Postural_DataC    83        1.4091907
#> Postural_DataC.txt.84  Postural_DataC    84        1.1520881
#> Postural_DataC.txt.85  Postural_DataC    85        2.5963749
#> Postural_DataC.txt.86  Postural_DataC    86        2.4004997
#> Postural_DataC.txt.87  Postural_DataC    87        3.6292277
#> Postural_DataC.txt.88  Postural_DataC    88        2.1870083
#> Postural_DataC.txt.89  Postural_DataC    89        1.4691579
#> Postural_DataC.txt.90  Postural_DataC    90        3.6420084
#> Postural_DataC.txt.91  Postural_DataC    91        3.2154352
#> Postural_DataC.txt.92  Postural_DataC    92        2.0718165
#> Postural_DataC.txt.93  Postural_DataC    93        1.4078880
#> Postural_DataC.txt.94  Postural_DataC    94        1.3534582
#> Postural_DataC.txt.95  Postural_DataC    95        1.3478050
#> Postural_DataC.txt.96  Postural_DataC    96        1.5890628
#> Postural_DataC.txt.97  Postural_DataC    97        2.9318607
#> Postural_DataC.txt.98  Postural_DataC    98        2.5094005
#> Postural_DataC.txt.99  Postural_DataC    99        1.7959432
#> Postural_DataC.txt.100 Postural_DataC   100        1.7977881
#> Postural_DataD.txt.1   Postural_DataD     1        2.0527509
#> Postural_DataD.txt.2   Postural_DataD     2        1.9357118
#> Postural_DataD.txt.3   Postural_DataD     3        2.1259591
#> Postural_DataD.txt.4   Postural_DataD     4        1.9279357
#> Postural_DataD.txt.5   Postural_DataD     5        1.5557037
#> Postural_DataD.txt.6   Postural_DataD     6        1.6844352
#> Postural_DataD.txt.7   Postural_DataD     7        1.3949547
#> Postural_DataD.txt.8   Postural_DataD     8        1.7824167
#> Postural_DataD.txt.9   Postural_DataD     9        1.3732626
#> Postural_DataD.txt.10  Postural_DataD    10        1.4678510
#> Postural_DataD.txt.11  Postural_DataD    11        1.6119868
#> Postural_DataD.txt.12  Postural_DataD    12        1.5901483
#> Postural_DataD.txt.13  Postural_DataD    13        1.6235027
#> Postural_DataD.txt.14  Postural_DataD    14        1.6598960
#> Postural_DataD.txt.15  Postural_DataD    15        1.7441205
#> Postural_DataD.txt.16  Postural_DataD    16        1.5040679
#> Postural_DataD.txt.17  Postural_DataD    17        1.4690987
#> Postural_DataD.txt.18  Postural_DataD    18        1.6682681
#> Postural_DataD.txt.19  Postural_DataD    19        1.6285053
#> Postural_DataD.txt.20  Postural_DataD    20        1.7438788
#> Postural_DataD.txt.21  Postural_DataD    21        1.5385810
#> Postural_DataD.txt.22  Postural_DataD    22        1.6697377
#> Postural_DataD.txt.23  Postural_DataD    23        1.7652677
#> Postural_DataD.txt.24  Postural_DataD    24        1.7706579
#> Postural_DataD.txt.25  Postural_DataD    25        1.4874324
#> Postural_DataD.txt.26  Postural_DataD    26        1.6188022
#> Postural_DataD.txt.27  Postural_DataD    27        1.5817829
#> Postural_DataD.txt.28  Postural_DataD    28        1.5588328
#> Postural_DataD.txt.29  Postural_DataD    29        1.5548303
#> Postural_DataD.txt.30  Postural_DataD    30        1.5882474
#> Postural_DataD.txt.31  Postural_DataD    31        1.8104936
#> Postural_DataD.txt.32  Postural_DataD    32        1.2685888
#> Postural_DataD.txt.33  Postural_DataD    33        1.4969159
#> Postural_DataD.txt.34  Postural_DataD    34        1.2842541
#> Postural_DataD.txt.35  Postural_DataD    35        1.4737792
#> Postural_DataD.txt.36  Postural_DataD    36        1.5511211
#> Postural_DataD.txt.37  Postural_DataD    37        2.2707187
#> Postural_DataD.txt.38  Postural_DataD    38        1.8117255
#> Postural_DataD.txt.39  Postural_DataD    39        2.0802555
#> Postural_DataD.txt.40  Postural_DataD    40        1.6600142
#> Postural_DataD.txt.41  Postural_DataD    41        1.7888933
#> Postural_DataD.txt.42  Postural_DataD    42        1.6312068
#> Postural_DataD.txt.43  Postural_DataD    43        1.6673795
#> Postural_DataD.txt.44  Postural_DataD    44        1.3621483
#> Postural_DataD.txt.45  Postural_DataD    45        1.3921760
#> Postural_DataD.txt.46  Postural_DataD    46        1.4643690
#> Postural_DataD.txt.47  Postural_DataD    47        1.9899453
#> Postural_DataD.txt.48  Postural_DataD    48        1.7363664
#> Postural_DataD.txt.49  Postural_DataD    49        1.4432395
#> Postural_DataD.txt.50  Postural_DataD    50        1.8147986
#> Postural_DataD.txt.51  Postural_DataD    51        1.5851557
#> Postural_DataD.txt.52  Postural_DataD    52        1.5416768
#> Postural_DataD.txt.53  Postural_DataD    53        1.4564300
#> Postural_DataD.txt.54  Postural_DataD    54        1.6479466
#> Postural_DataD.txt.55  Postural_DataD    55        1.4822581
#> Postural_DataD.txt.56  Postural_DataD    56        1.4479458
#> Postural_DataD.txt.57  Postural_DataD    57        1.6303925
#> Postural_DataD.txt.58  Postural_DataD    58        1.5835115
#> Postural_DataD.txt.59  Postural_DataD    59        1.6903310
#> Postural_DataD.txt.60  Postural_DataD    60        1.4307531
#> Postural_DataD.txt.61  Postural_DataD    61        1.6043797
#> Postural_DataD.txt.62  Postural_DataD    62        1.5406948
#> Postural_DataD.txt.63  Postural_DataD    63        1.5356977
#> Postural_DataD.txt.64  Postural_DataD    64        1.3039769
#> Postural_DataD.txt.65  Postural_DataD    65        1.4547596
#> Postural_DataD.txt.66  Postural_DataD    66        1.6780170
#> Postural_DataD.txt.67  Postural_DataD    67        1.6503491
#> Postural_DataD.txt.68  Postural_DataD    68        1.6263841
#> Postural_DataD.txt.69  Postural_DataD    69        1.4228013
#> Postural_DataD.txt.70  Postural_DataD    70        1.7738276
#> Postural_DataD.txt.71  Postural_DataD    71        1.6766823
#> Postural_DataD.txt.72  Postural_DataD    72        1.5550955
#> Postural_DataD.txt.73  Postural_DataD    73        1.5561150
#> Postural_DataD.txt.74  Postural_DataD    74        1.4429979
#> Postural_DataD.txt.75  Postural_DataD    75        1.7289006
#> Postural_DataD.txt.76  Postural_DataD    76        1.7650681
#> Postural_DataD.txt.77  Postural_DataD    77        1.5667184
#> Postural_DataD.txt.78  Postural_DataD    78        1.5487751
#> Postural_DataD.txt.79  Postural_DataD    79        1.3867417
#> Postural_DataD.txt.80  Postural_DataD    80        1.6970638
#> Postural_DataD.txt.81  Postural_DataD    81        1.9486499
#> Postural_DataD.txt.82  Postural_DataD    82        2.1105076
#> Postural_DataD.txt.83  Postural_DataD    83        1.5264844
#> Postural_DataD.txt.84  Postural_DataD    84        1.6079773
#> Postural_DataD.txt.85  Postural_DataD    85        1.6760482
#> Postural_DataD.txt.86  Postural_DataD    86        1.5394581
#> Postural_DataD.txt.87  Postural_DataD    87        1.3538656
#> Postural_DataD.txt.88  Postural_DataD    88        1.8084221
#> Postural_DataD.txt.89  Postural_DataD    89        1.6388548
#> Postural_DataD.txt.90  Postural_DataD    90        1.5015155
#> Postural_DataD.txt.91  Postural_DataD    91        1.5442945
#> Postural_DataD.txt.92  Postural_DataD    92        1.9364381
#> Postural_DataD.txt.93  Postural_DataD    93        1.5529080
#> Postural_DataD.txt.94  Postural_DataD    94        1.6497398
#> Postural_DataD.txt.95  Postural_DataD    95        1.5457456
#> Postural_DataD.txt.96  Postural_DataD    96        1.5339518
#> Postural_DataD.txt.97  Postural_DataD    97        1.4764462
#> Postural_DataD.txt.98  Postural_DataD    98        1.3772210
#> Postural_DataD.txt.99  Postural_DataD    99        1.3816914
#> Postural_DataD.txt.100 Postural_DataD   100        1.6250370
#> Postural_DataE.txt.1   Postural_DataE     1        1.5691013
#> Postural_DataE.txt.2   Postural_DataE     2        1.5787095
#> Postural_DataE.txt.3   Postural_DataE     3        1.4710667
#> Postural_DataE.txt.4   Postural_DataE     4        1.7043725
#> Postural_DataE.txt.5   Postural_DataE     5        1.5157870
#> Postural_DataE.txt.6   Postural_DataE     6        1.3094516
#> Postural_DataE.txt.7   Postural_DataE     7        1.5881714
#> Postural_DataE.txt.8   Postural_DataE     8        1.6554832
#> Postural_DataE.txt.9   Postural_DataE     9        1.6202244
#> Postural_DataE.txt.10  Postural_DataE    10        1.4673073
#> Postural_DataE.txt.11  Postural_DataE    11        1.5213312
#> Postural_DataE.txt.12  Postural_DataE    12        1.3254693
#> Postural_DataE.txt.13  Postural_DataE    13        1.6708960
#> Postural_DataE.txt.14  Postural_DataE    14        1.5636369
#> Postural_DataE.txt.15  Postural_DataE    15        1.6102425
#> Postural_DataE.txt.16  Postural_DataE    16        1.3322348
#> Postural_DataE.txt.17  Postural_DataE    17        1.3507934
#> Postural_DataE.txt.18  Postural_DataE    18        1.4873592
#> Postural_DataE.txt.19  Postural_DataE    19        1.5431385
#> Postural_DataE.txt.20  Postural_DataE    20        1.9384780
#> Postural_DataE.txt.21  Postural_DataE    21        1.4126408
#> Postural_DataE.txt.22  Postural_DataE    22        1.6829017
#> Postural_DataE.txt.23  Postural_DataE    23        1.5927613
#> Postural_DataE.txt.24  Postural_DataE    24        1.6189108
#> Postural_DataE.txt.25  Postural_DataE    25        1.6022942
#> Postural_DataE.txt.26  Postural_DataE    26        1.5881430
#> Postural_DataE.txt.27  Postural_DataE    27        1.5100981
#> Postural_DataE.txt.28  Postural_DataE    28        1.6071907
#> Postural_DataE.txt.29  Postural_DataE    29        1.4969721
#> Postural_DataE.txt.30  Postural_DataE    30        1.3867227
#> Postural_DataE.txt.31  Postural_DataE    31        1.6279688
#> Postural_DataE.txt.32  Postural_DataE    32        1.3930148
#> Postural_DataE.txt.33  Postural_DataE    33        1.7300183
#> Postural_DataE.txt.34  Postural_DataE    34        1.7830789
#> Postural_DataE.txt.35  Postural_DataE    35        1.6346477
#> Postural_DataE.txt.36  Postural_DataE    36        1.5376778
#> Postural_DataE.txt.37  Postural_DataE    37        1.6825636
#> Postural_DataE.txt.38  Postural_DataE    38        1.5515622
#> Postural_DataE.txt.39  Postural_DataE    39        1.6550853
#> Postural_DataE.txt.40  Postural_DataE    40        1.8014632
#> Postural_DataE.txt.41  Postural_DataE    41        1.6303400
#> Postural_DataE.txt.42  Postural_DataE    42        1.5603144
#> Postural_DataE.txt.43  Postural_DataE    43        1.5259051
#> Postural_DataE.txt.44  Postural_DataE    44        1.5263318
#> Postural_DataE.txt.45  Postural_DataE    45        1.6065871
#> Postural_DataE.txt.46  Postural_DataE    46        1.7219050
#> Postural_DataE.txt.47  Postural_DataE    47        1.7043506
#> Postural_DataE.txt.48  Postural_DataE    48        1.7710648
#> Postural_DataE.txt.49  Postural_DataE    49        1.5685124
#> Postural_DataE.txt.50  Postural_DataE    50        1.3927142
#> Postural_DataE.txt.51  Postural_DataE    51        1.4347791
#> Postural_DataE.txt.52  Postural_DataE    52        2.1311967
#> Postural_DataE.txt.53  Postural_DataE    53        1.5259943
#> Postural_DataE.txt.54  Postural_DataE    54        1.7314072
#> Postural_DataE.txt.55  Postural_DataE    55        1.7625144
#> Postural_DataE.txt.56  Postural_DataE    56        1.5109756
#> Postural_DataE.txt.57  Postural_DataE    57        2.2948228
#> Postural_DataE.txt.58  Postural_DataE    58        2.2976227
#> Postural_DataE.txt.59  Postural_DataE    59        1.5939144
#> Postural_DataE.txt.60  Postural_DataE    60        1.7068135
#> Postural_DataE.txt.61  Postural_DataE    61        1.4788771
#> Postural_DataE.txt.62  Postural_DataE    62        1.3988049
#> Postural_DataE.txt.63  Postural_DataE    63        1.4048872
#> Postural_DataE.txt.64  Postural_DataE    64        1.4273539
#> Postural_DataE.txt.65  Postural_DataE    65        1.4025739
#> Postural_DataE.txt.66  Postural_DataE    66        1.5427039
#> Postural_DataE.txt.67  Postural_DataE    67        1.7252239
#> Postural_DataE.txt.68  Postural_DataE    68        1.5017312
#> Postural_DataE.txt.69  Postural_DataE    69        1.3883112
#> Postural_DataE.txt.70  Postural_DataE    70        1.7237205
#> Postural_DataE.txt.71  Postural_DataE    71        1.7922230
#> Postural_DataE.txt.72  Postural_DataE    72        1.6495074
#> Postural_DataE.txt.73  Postural_DataE    73        1.8214466
#> Postural_DataE.txt.74  Postural_DataE    74        1.6965826
#> Postural_DataE.txt.75  Postural_DataE    75        1.5183882
#> Postural_DataE.txt.76  Postural_DataE    76        1.5859928
#> Postural_DataE.txt.77  Postural_DataE    77        1.6030272
#> Postural_DataE.txt.78  Postural_DataE    78        1.6864139
#> Postural_DataE.txt.79  Postural_DataE    79        1.5845667
#> Postural_DataE.txt.80  Postural_DataE    80        1.5477161
#> Postural_DataE.txt.81  Postural_DataE    81        1.7028350
#> Postural_DataE.txt.82  Postural_DataE    82        2.3325690
#> Postural_DataE.txt.83  Postural_DataE    83        1.9612556
#> Postural_DataE.txt.84  Postural_DataE    84        2.4078040
#> Postural_DataE.txt.85  Postural_DataE    85        2.5532938
#> Postural_DataE.txt.86  Postural_DataE    86        2.0687071
#> Postural_DataE.txt.87  Postural_DataE    87        1.8286529
#> Postural_DataE.txt.88  Postural_DataE    88        1.7105725
#> Postural_DataE.txt.89  Postural_DataE    89        1.6923481
#> Postural_DataE.txt.90  Postural_DataE    90        1.6811712
#> Postural_DataE.txt.91  Postural_DataE    91        1.7613695
#> Postural_DataE.txt.92  Postural_DataE    92        2.0414001
#> Postural_DataE.txt.93  Postural_DataE    93        2.0674678
#> Postural_DataE.txt.94  Postural_DataE    94        1.8467204
#> Postural_DataE.txt.95  Postural_DataE    95        1.9764743
#> Postural_DataE.txt.96  Postural_DataE    96        1.8852092
#> Postural_DataE.txt.97  Postural_DataE    97        1.5783082
#> Postural_DataE.txt.98  Postural_DataE    98        1.5014515
#> Postural_DataE.txt.99  Postural_DataE    99        1.4925955
#> Postural_DataE.txt.100 Postural_DataE   100        1.4007763
#> Postural_DataF.txt.1   Postural_DataF     1        1.5952567
#> Postural_DataF.txt.2   Postural_DataF     2        2.0084787
#> Postural_DataF.txt.3   Postural_DataF     3        2.0680212
#> Postural_DataF.txt.4   Postural_DataF     4        2.0159859
#> Postural_DataF.txt.5   Postural_DataF     5        1.9704170
#> Postural_DataF.txt.6   Postural_DataF     6        2.1852298
#> Postural_DataF.txt.7   Postural_DataF     7        1.8381908
#> Postural_DataF.txt.8   Postural_DataF     8        1.8017421
#> Postural_DataF.txt.9   Postural_DataF     9        1.7439671
#> Postural_DataF.txt.10  Postural_DataF    10        1.5618304
#> Postural_DataF.txt.11  Postural_DataF    11        1.6297856
#> Postural_DataF.txt.12  Postural_DataF    12        1.7031536
#> Postural_DataF.txt.13  Postural_DataF    13        1.4691228
#> Postural_DataF.txt.14  Postural_DataF    14        1.3460970
#> Postural_DataF.txt.15  Postural_DataF    15        1.7085884
#> Postural_DataF.txt.16  Postural_DataF    16        1.5349116
#> Postural_DataF.txt.17  Postural_DataF    17        1.7519758
#> Postural_DataF.txt.18  Postural_DataF    18        1.5400262
#> Postural_DataF.txt.19  Postural_DataF    19        1.6567330
#> Postural_DataF.txt.20  Postural_DataF    20        1.6356364
#> Postural_DataF.txt.21  Postural_DataF    21        1.3649455
#> Postural_DataF.txt.22  Postural_DataF    22        1.3346984
#> Postural_DataF.txt.23  Postural_DataF    23        1.4924634
#> Postural_DataF.txt.24  Postural_DataF    24        1.3969626
#> Postural_DataF.txt.25  Postural_DataF    25        1.4344384
#> Postural_DataF.txt.26  Postural_DataF    26        1.7627528
#> Postural_DataF.txt.27  Postural_DataF    27        1.4822280
#> Postural_DataF.txt.28  Postural_DataF    28        1.4861272
#> Postural_DataF.txt.29  Postural_DataF    29        1.4925624
#> Postural_DataF.txt.30  Postural_DataF    30        1.5501082
#> Postural_DataF.txt.31  Postural_DataF    31        1.5246271
#> Postural_DataF.txt.32  Postural_DataF    32        1.4997147
#> Postural_DataF.txt.33  Postural_DataF    33        1.5079677
#> Postural_DataF.txt.34  Postural_DataF    34        1.6950881
#> Postural_DataF.txt.35  Postural_DataF    35        1.5905226
#> Postural_DataF.txt.36  Postural_DataF    36        1.4725947
#> Postural_DataF.txt.37  Postural_DataF    37        1.2947929
#> Postural_DataF.txt.38  Postural_DataF    38        1.6293214
#> Postural_DataF.txt.39  Postural_DataF    39        1.4153533
#> Postural_DataF.txt.40  Postural_DataF    40        1.6504723
#> Postural_DataF.txt.41  Postural_DataF    41        1.4113570
#> Postural_DataF.txt.42  Postural_DataF    42        1.7834995
#> Postural_DataF.txt.43  Postural_DataF    43        1.4578171
#> Postural_DataF.txt.44  Postural_DataF    44        1.5969950
#> Postural_DataF.txt.45  Postural_DataF    45        1.4903426
#> Postural_DataF.txt.46  Postural_DataF    46        1.5081506
#> Postural_DataF.txt.47  Postural_DataF    47        1.4951076
#> Postural_DataF.txt.48  Postural_DataF    48        1.5426842
#> Postural_DataF.txt.49  Postural_DataF    49        1.7639463
#> Postural_DataF.txt.50  Postural_DataF    50        1.5378116
#> Postural_DataF.txt.51  Postural_DataF    51        1.8007653
#> Postural_DataF.txt.52  Postural_DataF    52        1.6771825
#> Postural_DataF.txt.53  Postural_DataF    53        1.9052654
#> Postural_DataF.txt.54  Postural_DataF    54        1.4938874
#> Postural_DataF.txt.55  Postural_DataF    55        1.5991339
#> Postural_DataF.txt.56  Postural_DataF    56        1.6943725
#> Postural_DataF.txt.57  Postural_DataF    57        1.5518777
#> Postural_DataF.txt.58  Postural_DataF    58        1.6724590
#> Postural_DataF.txt.59  Postural_DataF    59        1.5186014
#> Postural_DataF.txt.60  Postural_DataF    60        1.6586406
#> Postural_DataF.txt.61  Postural_DataF    61        1.6236262
#> Postural_DataF.txt.62  Postural_DataF    62        1.7437301
#> Postural_DataF.txt.63  Postural_DataF    63        1.7304747
#> Postural_DataF.txt.64  Postural_DataF    64        1.7581892
#> Postural_DataF.txt.65  Postural_DataF    65        1.5148164
#> Postural_DataF.txt.66  Postural_DataF    66        1.6801743
#> Postural_DataF.txt.67  Postural_DataF    67        1.7765071
#> Postural_DataF.txt.68  Postural_DataF    68        1.7119578
#> Postural_DataF.txt.69  Postural_DataF    69        1.3986555
#> Postural_DataF.txt.70  Postural_DataF    70        1.8437034
#> Postural_DataF.txt.71  Postural_DataF    71        1.7115469
#> Postural_DataF.txt.72  Postural_DataF    72        1.6999754
#> Postural_DataF.txt.73  Postural_DataF    73        1.5246759
#> Postural_DataF.txt.74  Postural_DataF    74        1.4502448
#> Postural_DataF.txt.75  Postural_DataF    75        1.8793287
#> Postural_DataF.txt.76  Postural_DataF    76        1.5113943
#> Postural_DataF.txt.77  Postural_DataF    77        1.4361086
#> Postural_DataF.txt.78  Postural_DataF    78        1.7008644
#> Postural_DataF.txt.79  Postural_DataF    79        1.2496002
#> Postural_DataF.txt.80  Postural_DataF    80        1.6197295
#> Postural_DataF.txt.81  Postural_DataF    81        1.6614708
#> Postural_DataF.txt.82  Postural_DataF    82        1.6009147
#> Postural_DataF.txt.83  Postural_DataF    83        1.6480290
#> Postural_DataF.txt.84  Postural_DataF    84        2.1248011
#> Postural_DataF.txt.85  Postural_DataF    85        1.5852633
#> Postural_DataF.txt.86  Postural_DataF    86        1.7102562
#> Postural_DataF.txt.87  Postural_DataF    87        1.6490652
#> Postural_DataF.txt.88  Postural_DataF    88        1.7446713
#> Postural_DataF.txt.89  Postural_DataF    89        1.7782024
#> Postural_DataF.txt.90  Postural_DataF    90        1.5592692
#> Postural_DataF.txt.91  Postural_DataF    91        1.5869012
#> Postural_DataF.txt.92  Postural_DataF    92        1.5420915
#> Postural_DataF.txt.93  Postural_DataF    93        1.5349243
#> Postural_DataF.txt.94  Postural_DataF    94        1.5746787
#> Postural_DataF.txt.95  Postural_DataF    95        1.4017731
#> Postural_DataF.txt.96  Postural_DataF    96        1.4233727
#> Postural_DataF.txt.97  Postural_DataF    97        1.4395912
#> Postural_DataF.txt.98  Postural_DataF    98        1.4281614
#> Postural_DataF.txt.99  Postural_DataF    99        1.5034584
#> Postural_DataF.txt.100 Postural_DataF   100        1.5446302