Create a "spaghetti plot" from your postural data.

This function creates a spaghetti plot - a plot that illustrates the CoP displacements stored in a data frame. The data frame should contain minimally a column storing the time course, a column storing CoP-X values, and a column storing CoP-Y values.

Usage

Spaghett(
  df,
  time_col = "Time",
  copx_col = "CoP_X",
  copy_col = "CoP_Y",
  Title = "Spaghetti plot",
  xlab = "Mean CoP_X",
  ylab = "Mean CoP_Y"
)

Arguments

df

dataframe containing the data you want to plot

time_col

column from dataframe that indicates the time course of postural measurements

copx_col

column from dataframe that contains the CoP-X displacements

copy_col

column from dataframe that contains the CoP-Y displacements

Title

Character string specifying a title for your plot (optionnal)

xlab

Character string specifying the x-axis name (optional)

ylab

Character string specifying the y-axis name (optional)

Value

a graphical object

Examples

data(Postural_DataA) # Load a session to plot
Spaghett(Postural_DataA, time_col = "Time", copx_col = "CoP_X", copy_col = "CoP_Y")
#> Error in Spaghett(Postural_DataA, time_col = "Time", copx_col = "CoP_X",     copy_col = "CoP_Y"): One or more specified columns do not exist in the dataframe.